+++ co_fact
@@ -85,8 +85,8 @@ text_buffer_size 2048
 
 # -- Lua load -- #
 lua_load fact.lua
-lua_draw_hook_pre main
-lua_startup_hook setup
+lua_draw_hook_pre main_fact
+lua_startup_hook setup_fact
 
 TEXT
 ${execi 100 Scripts/fact.py}
