+++ co_main
@@ -84,8 +84,8 @@ use_spacer right
 text_buffer_size 2048
 
 # -- Lua load -- #
-lua_load main.lua
-lua_draw_hook_pre main
-lua_startup_hook setup
+lua_load system.lua
+lua_draw_hook_pre main_system
+lua_startup_hook setup_system
 
 TEXT
