+++ co_quote
@@ -85,8 +85,8 @@ text_buffer_size 2048
 
 # -- Lua load -- #
 lua_load quote.lua
-lua_draw_hook_pre main
-lua_startup_hook setup
+lua_draw_hook_pre main_quote
+lua_startup_hook setup_quote
 
 TEXT
 ${execi 100 Scripts/quote.py}
