+++ co_weather
@@ -85,8 +85,8 @@ text_buffer_size 2048
 
 # -- Lua load -- #
 lua_load weather.lua
-lua_draw_hook_pre main
-lua_startup_hook setup
+lua_draw_hook_pre main_weather
+lua_startup_hook setup_weather
 
 TEXT
 ${execi 100 Scripts/weather.py}
